move getScrollParent to helpers
